refactor: seperate interceptor code
